Fixed decision if archive needs to be fetched
